Encapsulation 2014
This Java 6 programme hides all my variables from the rest of the
programme. I managed this by making my variables ints all Private and
the class that they are in a Public class so that the public methods in
my other public class can see them and use them ok.